requirements.txt
torch
torchvision
opencv-python
numpy
pyyaml
matplotlib
tqdm
pillow